Reorganize Editor components
