+++ src/renderer/screens/Editor/index.js
@@ -1,0 +1,3 @@
+import Editor from "./Editor";
+
+export default Editor;
